requirements.txt
customtkinter
tk
pillow
pydub
SpeechRecognition
PyAudio
lxml
python-docx
openai
requests